backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-filter==23.1
djoser==2.1.0
Pillow==9.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.3